KW-179/refactor: area service -> hospital service
