+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'area-service'
+rootProject.name = 'hospital-service'
