--- src/main/java/com/doubleo/areaservice/AreaServiceApplication.java
@@ -1,12 +1,0 @@
-package com.doubleo.areaservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AreaServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AreaServiceApplication.class, args);
-    }
-}
