--- src/main/java/com/doubleo/areaservice/domain/area/domain/AreaCategory.java
@@ -1,7 +1,0 @@
-package com.doubleo.areaservice.domain.area.domain;
-
-public enum AreaCategory {
-    OFFICE,
-    MEETING_ROOM,
-    OTHER
-}
