--- src/main/java/com/doubleo/areaservice/domain/area/domain/EnterStatus.java
@@ -1,23 +1,0 @@
-package com.doubleo.areaservice.domain.area.domain;
-
-import com.doubleo.areaservice.domain.common.model.BaseTimeEntity;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor
-public class EnterStatus extends BaseTimeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long enterStatusId;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "area_id", nullable = false)
-    private Area area;
-
-    @Column(nullable = false)
-    private Long employeeId;
-}
