--- src/main/java/com/doubleo/areaservice/domain/area/dto/response/BuildingInfoResponse.java
@@ -1,18 +1,0 @@
-package com.doubleo.areaservice.domain.area.dto.response;
-
-import com.doubleo.areaservice.domain.area.domain.Building;
-import lombok.Getter;
-
-@Getter
-public class BuildingInfoResponse {
-
-    private Long buildingId;
-    private String buildingName;
-
-    public static BuildingInfoResponse from(Building building) {
-        BuildingInfoResponse response = new BuildingInfoResponse();
-        response.buildingId = building.getBuildingId();
-        response.buildingName = building.getBuildingName();
-        return response;
-    }
-}
