--- src/main/java/com/doubleo/areaservice/domain/area/repository/BuildingRepository.java
@@ -1,6 +1,0 @@
-package com.doubleo.areaservice.domain.area.repository;
-
-import com.doubleo.areaservice.domain.area.domain.Building;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BuildingRepository extends JpaRepository<Building, Long> {}
