--- src/main/java/com/doubleo/areaservice/global/config/swagger/SwaggerConfig.java
@@ -1,20 +1,0 @@
-package com.doubleo.areaservice.global.config.swagger;
-
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class SwaggerConfig {
-
-    @Bean
-    public OpenAPI openAPI() {
-        return new OpenAPI()
-                .info(
-                        new Info()
-                                .title("Area Service API") // 서비스 이름
-                                .description("출입구역 API 명세서")
-                                .version("v1.0.0"));
-    }
-}
