--- src/main/java/com/doubleo/areaservice/global/exception/CommonException.java
@@ -1,15 +1,0 @@
-package com.doubleo.areaservice.global.exception;
-
-import com.doubleo.areaservice.global.exception.errorcode.BaseErrorCode;
-import lombok.Getter;
-
-@Getter
-public class CommonException extends RuntimeException {
-
-    private final BaseErrorCode errorCode;
-
-    public CommonException(BaseErrorCode errorCode) {
-        super(errorCode.getMessage());
-        this.errorCode = errorCode;
-    }
-}
