--- src/main/java/com/doubleo/areaservice/global/exception/ErrorResponse.java
@@ -1,7 +1,0 @@
-package com.doubleo.areaservice.global.exception;
-
-public record ErrorResponse(String errorClassName, String message) {
-    public static ErrorResponse of(String errorClassName, String message) {
-        return new ErrorResponse(errorClassName, message);
-    }
-}
