--- src/main/java/com/doubleo/areaservice/global/exception/errorcode/BaseErrorCode.java
@@ -1,11 +1,0 @@
-package com.doubleo.areaservice.global.exception.errorcode;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseErrorCode {
-    HttpStatus getHttpStatus();
-
-    String getMessage();
-
-    String errorClassName();
-}
