--- src/main/java/com/doubleo/areaservice/infra/config/jpa/JpaAuditingConfig.java
@@ -1,8 +1,0 @@
-package com.doubleo.areaservice.infra.config.jpa;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaAuditingConfig {}
