+++ src/main/java/com/doubleo/hospitalservice/HospitalServiceApplication.java
@@ -1,0 +1,12 @@
+package com.doubleo.hospitalservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HospitalServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(HospitalServiceApplication.class, args);
+    }
+}
