+++ src/main/java/com/doubleo/hospitalservice/domain/hospital/domain/Hospital.java
@@ -1,0 +1,13 @@
+package com.doubleo.hospitalservice.domain.hospital.domain;
+
+import com.doubleo.hospitalservice.domain.common.model.BaseTimeEntity;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Hospital extends BaseTimeEntity {
+    @Id private Long id;
+}
