+++ src/main/java/com/doubleo/hospitalservice/global/exception/CommonException.java
@@ -1,0 +1,15 @@
+package com.doubleo.hospitalservice.global.exception;
+
+import com.doubleo.hospitalservice.global.exception.errorcode.BaseErrorCode;
+import lombok.Getter;
+
+@Getter
+public class CommonException extends RuntimeException {
+
+    private final BaseErrorCode errorCode;
+
+    public CommonException(BaseErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.errorCode = errorCode;
+    }
+}
