+++ src/main/java/com/doubleo/hospitalservice/global/exception/ErrorResponse.java
@@ -1,0 +1,7 @@
+package com.doubleo.hospitalservice.global.exception;
+
+public record ErrorResponse(String errorClassName, String message) {
+    public static ErrorResponse of(String errorClassName, String message) {
+        return new ErrorResponse(errorClassName, message);
+    }
+}
