+++ src/main/java/com/doubleo/hospitalservice/global/exception/errorcode/BaseErrorCode.java
@@ -1,0 +1,11 @@
+package com.doubleo.hospitalservice.global.exception.errorcode;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseErrorCode {
+    HttpStatus getHttpStatus();
+
+    String getMessage();
+
+    String errorClassName();
+}
