+++ src/main/java/com/doubleo/hospitalservice/global/exception/errorcode/HospitalErrorCode.java
@@ -1,0 +1,20 @@
+package com.doubleo.hospitalservice.global.exception.errorcode;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+@AllArgsConstructor
+public enum HospitalErrorCode implements BaseErrorCode {
+    HOSPITAL_NOT_FOUND(HttpStatus.NOT_FOUND, "병원 정보를 찾을 수 없습니다."),
+    ;
+
+    private final HttpStatus httpStatus;
+    private final String message;
+
+    @Override
+    public String errorClassName() {
+        return this.name();
+    }
+}
