+++ src/main/java/com/doubleo/hospitalservice/infra/config/jpa/JpaAuditingConfig.java
@@ -1,0 +1,8 @@
+package com.doubleo.hospitalservice.infra.config.jpa;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaAuditingConfig {}
