+++ src/main/resources/application.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: area-service
+    name: hospital-service
 
   config:
     import: optional:configserver:${CONFIG_SERVER_URL}
