--- src/test/java/com/doubleo/areaservice/AreaServiceApplicationTests.java
@@ -1,11 +1,0 @@
-package com.doubleo.areaservice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AreaServiceApplicationTests {
-
-    @Test
-    void contextLoads() {}
-}
