+++ src/test/java/com/doubleo/hospitalservice/HospitalServiceApplicationTests.java
@@ -1,0 +1,11 @@
+package com.doubleo.hospitalservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HospitalServiceApplicationTests {
+
+    @Test
+    void contextLoads() {}
+}
